Move import logic out of dlalog because this doalog only called in DragFileSubScreen So don't need to collect all logic in that place.
