+++ osu.Game.Rulesets.Karaoke/Edit/ImportLyric/DragFile/ImportLyricDialog.cs
@@ -15,9 +15,6 @@ namespace osu.Game.Rulesets.Karaoke.Edit.ImportLyric.DragFile
     public class ImportLyricDialog : PopupDialog
     {
         [Resolved]
-        private ImportLyricManager importManager { get; set; }
-
-        [Resolved]
         private DialogOverlay dialogOverlay { get; set; }
 
         public ImportLyricDialog(FileInfo info, Action<bool> resetAction = null)
@@ -31,96 +28,14 @@ namespace osu.Game.Rulesets.Karaoke.Edit.ImportLyric.DragFile
                 new PopupDialogOkButton
                 {
                     Text = @"Yes. Go for it.",
-                    Action = () =>
-                    {
-                        var success = processImport(info);
-                        resetAction?.Invoke(success);
-                    }
+                    Action = () => resetAction?.Invoke(true),
                 },
                 new PopupDialogCancelButton
                 {
                     Text = @"No! Abort mission!",
+                    Action = () => resetAction?.Invoke(false),
                 },
             };
         }
-
-        private bool processImport(FileInfo info)
-        {
-            try
-            {
-                // Check file is exist
-                if (!info.Exists)
-                {
-                    dialogOverlay.Push(createFileNotFoundDialog());
-                    return false;
-                }
-
-                // Check format is match
-                if (!ImportLyricManager.LyricFormatExtensions.Contains(info.Extension))
-                {
-                    dialogOverlay.Push(createFormatNotMatchDialog());
-                    return false;
-                }
-
-                importManager.ImportLrcFile(info);
-                dialogOverlay.Push(new OkPopupDialog
-                {
-                    Icon = FontAwesome.Regular.CheckCircle,
-                    HeaderText = @"Import success",
-                    BodyText = "Lyrics has been imported."
-                });
-                return true;
-            }
-            catch (Exception ex)
-            {
-                switch (ex)
-                {
-                    case FileNotFoundException _:
-                        dialogOverlay.Push(createFileNotFoundDialog());
-                        break;
-
-                    case FileLoadException loadException:
-                        dialogOverlay.Push(createLoadExceptionDialog(loadException));
-                        break;
-                    default:
-                        dialogOverlay.Push(createUnknownExceptionDialog());
-                        break;
-                }
-
-                return false;
-            }
-        }
-
-        private PopupDialog createFileNotFoundDialog()
-            => new OkPopupDialog
-            {
-                Icon = FontAwesome.Regular.QuestionCircle,
-                HeaderText = "Seems file is not exist",
-                BodyText = $"Drag the file then drop again.",
-            };
-
-        private PopupDialog createFormatNotMatchDialog()
-            => new OkPopupDialog
-            {
-                Icon = FontAwesome.Solid.ExclamationTriangle,
-                HeaderText = "This type of file is not supported",
-                BodyText = $"May sure this type of file is supported.",
-            };
-
-        private PopupDialog createLoadExceptionDialog(FileLoadException loadException)
-            => new OkPopupDialog
-            {
-                Icon = FontAwesome.Solid.Bug,
-                HeaderText = @"File loading error",
-                BodyText = loadException.Message,
-            };
-
-        private PopupDialog createUnknownExceptionDialog()
-            => new OkPopupDialog
-            {
-                Icon = FontAwesome.Solid.Bug,
-                HeaderText = @"Unknown error",
-                BodyText = @"Unknown error QAQa."
-            };
     }
 }
